refactor: elliminate shared type code thats being removed from mongodb-extended-json
