+++ lib/index.js
@@ -1,13 +1,12 @@
 'use strict';
 
 const EJSON = require('mongodb-extended-json');
+const stringify = require('./stringify');
 const bson = require('bson');
 const ms = require('ms');
 const { SaferEval } = require('safer-eval');
-const toJavascriptString = require('javascript-stringify');
 const _ = require('lodash');
 const queryLanguage = require('mongodb-language-model');
-const getMongoDBType = require('mongodb-extended-json/lib/types').type;
 const debug = require('debug')('mongodb-query-parser');
 const { COLLATION_OPTIONS } = require('./constants');
 
@@ -19,74 +18,6 @@ const DEFAULT_PROJECT = null;
 const DEFAULT_COLLATION = null;
 const DEFAULT_MAX_TIME_MS = ms('5 seconds');
 const QUERY_PROPERTIES = ['filter', 'project', 'sort', 'skip', 'limit'];
-
-const BSON_TO_JS_STRING = {
-  Code: function(v) {
-    if (v.scope) {
-      return `Code('${v.code}',${JSON.stringify(v.scope)})`;
-    }
-    return `Code('${v.code}')`;
-  },
-  ObjectID: function(v) {
-    return `ObjectId('${v.toHexString()}')`;
-  },
-  ObjectId: function(v) {
-    return `ObjectId('${v.toHexString()}')`;
-  },
-  Binary: function(v) {
-    const subType = v.sub_type;
-    if (subType === 4) {
-      return `UUID("${v.buffer.toString('hex')}")`;
-    }
-    return `BinData(${subType.toString(16)}, '${v.buffer.toString('base64')}')`;
-  },
-  DBRef: function(v) {
-    return `DBRef('${v.namespace}', '${v.oid}')`;
-  },
-  Timestamp: function(v) {
-    return `Timestamp(${v.low}, ${v.high})`;
-  },
-  Long: function(v) {
-    return `NumberLong(${v.toString()})`;
-  },
-  Decimal128: function(v) {
-    return `NumberDecimal('${v.toString()}')`;
-  },
-  Int32: function(v) {
-    return `NumberInt{'${v.toString()}')`;
-  },
-  MaxKey: function() {
-    return 'MaxKey()';
-  },
-  MinKey: function() {
-    return 'MinKey()';
-  },
-  Date: function(v) {
-    return `ISODate('${v.toISOString()}')`;
-  },
-  ISODate: function(v) {
-    return `ISODate('${v.toISOString()}')`;
-  },
-  RegExp: function(v) {
-    let o = '';
-    let hasOptions = false;
-
-    if (v.global) {
-      hasOptions = true;
-      o += 'g';
-    }
-    if (v.ignoreCase) {
-      hasOptions = true;
-      o += 'i';
-    }
-    if (v.multiline) {
-      hasOptions = true;
-      o += 'm';
-    }
-
-    return `RegExp('${v.source}'${hasOptions ? `, ${o}` : ''})`;
-  }
-};
 
 const FILTER_SANDBOX = {
   RegExp: RegExp,
@@ -110,20 +41,30 @@ const FILTER_SANDBOX = {
   NumberInt: function(s) {
     return parseInt(s, 10);
   },
-  Long: bson.Long,
+  Long: function(low, high) {
+    return new bson.Long(low, high);
+  },
   NumberLong: function(v) {
     return bson.Long.fromNumber(v);
   },
   Int64: bson.Long,
   Map: bson.Map,
-  MaxKey: bson.MaxKey,
-  MinKey: bson.MinKey,
-  ObjectID: bson.ObjectID,
-  ObjectId: bson.ObjectID,
-  Symbol: bson.Symbol,
-  Timestamp: function(low, high) {
-    return new bson.Timestamp(low, high);
+  MaxKey: function() {
+    return new bson.MaxKey();
   },
+  MinKey: function() {
+    return new bson.MinKey();
+  },
+  ObjectID: function(i) {
+    return new bson.ObjectID(i);
+  },
+  ObjectId: function(i) {
+    return new bson.ObjectID(i);
+  },
+  Symbol: function(i) {
+    return new bson.Symbol(i);
+  },
+  Timestamp: bson.Timestamp,
   ISODate: function(s) {
     return s === undefined ? new Date() : new Date(s);
   },
@@ -153,6 +94,9 @@ function isNumberValid(input) {
 
 function executeJavascript(input) {
   'use strict';
+  if (input.indexOf('clearImmediate.constructor') > -1) {
+    throw new TypeError('Not allowed to breakout');
+  }
   return SANDBOX.runInContext(input);
 }
 
@@ -379,27 +323,8 @@ module.exports.validate = function(what, input) {
   return validator(input);
 };
 
-function toJSString(obj, ind) {
-  return toJavascriptString(
-    obj,
-    function(value, indent, stringify) {
-      const toJs = BSON_TO_JS_STRING[getMongoDBType(value)];
-      if (!toJs) {
-        return stringify(value);
-      }
-      return toJs(value);
-    },
-    ind || ' '
-  );
-}
-
-module.exports.toJSString = toJSString;
-
-module.exports.stringify = function(obj) {
-  return toJSString(obj)
-    .replace(/ ?\n ? ?/g, '')
-    .replace(/ {2,}/g, ' ');
-};
+module.exports.toJSString = stringify.toJSString;
+module.exports.stringify = stringify.stringify;
 
 module.exports.QUERY_PROPERTIES = QUERY_PROPERTIES;
 module.exports.DEFAULT_FILTER = DEFAULT_FILTER;
